seo: make games and GFS model pages indexable, fix blog images and dedup
- Split app/games/[slug]/page.tsx into a server component with
  generateMetadata + generateStaticParams; interactive UI moved to
  game-detail-client.tsx. Games now render as SSG with per-game titles
  and canonicals, and are added to the sitemap via server-side Supabase
  query.
- New app/gfs-model/[region]/[run]/layout.tsx exports generateMetadata +
  generateStaticParams for 24 region/run combinations so Googlebot sees
  real per-run metadata without JS.
- Blog images converted to next/image: hero (priority-loaded for LCP),
  markdown renderer, and blog-index featured card. `unoptimized` flag
  used because heroImages are served by our own /api/og endpoint.
- Blog tag filter (/blog?tag=...) now emits robots: noindex + canonical
  to /blog so crawl budget concentrates on the canonical index and
  individual posts.
- Updated tests that were coupled to the old layout: sitemap now async
  (3 files) and the postMessage-origin check points at the new client
  component.
